requirements.txt
streamlit
google-generativeai
opencv-python-headless==4.8.0.76
Pillow
numpy
mediapipe
